rename cmp
