+++ init.lua
@@ -2,7 +2,7 @@ require "user.options"
 require "user.keymaps"
 require "user.plugins"
 require "user.colorscheme"
-require "user.cmp"
+require "user.completion"
 require "user.lsp"
 require "user.telescope"
 require "user.gitsigns"
