+++ lua/user/completion.lua
